@@ -1,15 +1,9 @@
 ﻿using MrMeeseeks.DIE;
-using MrMeeseeks.DIE.SampleChild;
-using SampleChild;
+using MrMeeseeks.DIE.Sample;
 
-[assembly:Spy(typeof(IPublicTypeReport), typeof(IInternalTypeReport))]
+//[assembly:Spy(typeof(IPublicTypeReport), typeof(IInternalTypeReport))]
 [assembly:SingleInstance(typeof(ISingleInstance))]
 [assembly:ScopedInstance(typeof(IScopedInstance))]
 [assembly:ScopeRoot(typeof(IScopeRoot))]
 [assembly:Transient(typeof(ITransient))]
-
-namespace MrMeeseeks.DIE.Sample;
-
-internal partial class Container : IContainer<IContext>
-{
-}
+[assembly:Decorator(typeof(IDecorator<>))]
