@@ -1,6 +1,7 @@
-namespace MrMeeseeks.DIE.SampleChild;
+namespace MrMeeseeks.DIE.Sample;
 
 public interface ISingleInstance { }
 public interface IScopedInstance { }
 public interface IScopeRoot { }
 public interface ITransient { }
+public interface IDecorator<T> { }
